Update contractor impeeding time estimates
Based on feedback from ATC 138 PRP. Updated to simulate mediate times
based on number of damaged systems and whether or not contractors are on
retainer. Defaults are defined within the function and are not
currently user inputs. Impedint times are simulated assuming a truncated
lognormal distribution. Currently assumes all systems are correlated.